add order controller

+++ db/migrate/20130326133222_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration
   def change
     create_table :orders do |t|
       t.integer :user_id
-      t.decimal :total
+      t.integer :total
       
       t.timestamps
     end
